Rework the test suite - add upstream tests
This cleans up parts of the Test Suite and adds test data from the
less.js library. Of course LesserPHP does not pass all the tests and
probably never will.
Some tests have been completely disabled for now and might be deleted
later. Other tests have been marked to be skipped - these need to be
looked at individually to decide how to deal with the failures.